it is working well

+++ mean-course/backend/routes/posts.js
@@ -68,9 +68,15 @@ multer({storage: storage}).single("image"),
     imagePath: imagePath
   });
   Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
-    res.status(200).json({ message: "Update successful!"});
-  })
-});
+    if (result.nModified > 0) {
+      res.status(200).json({ message: "Update successful!"});
+    }
+    else {
+      res.status(401).json({ message: "Not authorized!"});
+    }
+  });
+}
+);
 
 
 router.get('',(req, res, next) => {
@@ -107,9 +113,14 @@ router.get("/:id", (req, res, next) => {
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
     console.log(result);
-    res.status(200).json({ message: "Post deleted!"});
+    if (result.n > 0) {
+      res.status(200).json({ message: "Deletion successful!"});
+    }
+    else {
+      res.status(401).json({ message: "Not authorized!"});
+    }
   });
 });
